requirements.txt
google.generativeai
python-dotenv
flask
fastapi[standard]
uvicorn
streamlit